services/user-service/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
pymongo==4.4.1
PyJWT==2.8.0
Werkzeug==2.3.6
prometheus-flask-exporter==0.22.4
pytest==7.4.0
requests==2.31.0